Sync image removals to new frame renderings
When we're removing an old image we must make sure it's still not
referenced in a rendering that's still ongoing. To do that add a frame
count so we can know in what frame we've removed the image and to only
really remove it from memory when we're already starting the rendering
of the next one.