Don't leak image buffers in the conrod image map
The previous code had a much more complex tracking of IDs. It seems
quite simple to just reuse the same id, replacing the buffer in the map.
The only thing this doesn't do is remove the last image from the map
when we suddenly don't have any image selected. Seems like a minor
optimization unless someone is dealing with incredibly large images in
which case they'd have other problems.